Add styling

+++ db/migrate/20231009162528_create_deals.rb
@@ -2,7 +2,7 @@ class CreateDeals < ActiveRecord::Migration[7.0]
   def change
     create_table :deals do |t|
       t.string :name
-      t.integer :amount
+      t.decimal :amount
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
